I removed this header file so that we didn't...
I removed this header file so that we didn't have duplicated code
between it and peopt.h.  PEopt.hpp contained DGM specific code that
wasn't appropriate for here.  For example, the extra namespace is
unnecessary in a single application.  Also, the names for the different
kinds of hessian approximations should not exist here, but in DGM.  The
peopt.h header does *not* implement any hessian approximations and
therefore should not elaborate on which approximations are possible.  In
the diff_operator header, we do define mechanisms to automatically
derive some of these operators, but this is separate from the code in
peopt.h or PEopt.hpp.