Removed the old PEopt.
Removed the old PEopt.hpp which is not subsummed by peopt.h and added a
sketch of documentation for the peopt namespace in peopt.h so that at
the very least, the overall namespace is documented.  Doxygen will not
pick up that the included files (such as core.h) are in the peopt
namespace so this is still an issue.  However, with the customer sprint
over, we now have time to try to resolve this.