I altered how we determine a new trust-region...
I altered how we determine a new trust-region radius.  Now, we don't
increase the size of the trust-region radius unless it was a limiting
factoring in the previous iteration.  In addition, when we have a bad
step and the length of the step is less than the trust-region radius, we
scale the trust-region radius based on the length of the step.  This
should hopefully eliminate a number of unnecessary iterations while the
trust-region radius is whiddled away.