I added the next generation of the...
I added the next generation of the optimization code to the v2
directory.  Eventually, this will replace the old code.  The biggest
change is that the organization and interface now different.
Specifically, we added the concept of an optimization state, which
simplifies both the diagnostics as well as the interfaces.  We also
change how we interact with the algebra with explicit vector space
packages.  We accomplish this through a template trick, but it really
amounts to an existential type.  Finally, this code should more
completely fill out missing routines such as the linesearch methods.