I added a myriad of tests to the code to help...
I added a myriad of tests to the code to help verify the validity of the
functions.  This includes a number of finite difference tests, but it
also includes some adjoint and symmetry tests as well.  Also, I noticed
a problem with the algebra that I hadn't run across before.
Essentially, in the past, I used the scal function in order to set a
vector to 0.  Basically, we scale the vector my zero to make it zero.
However, this doesn't always work.  Specifically, if the vector contains
some odd elements such as NaNs, scaling by zero doesn't really set the
vector to zero.  Mathematically, this makes sense since we've
essentially broken the vector space structure.  Practically, we need
some way to handle this elegantly.  At the moment, the easy fix is to
require an actual zero function that will set a vector to 0.  This
basically clears out all bad elements and allows us to start from
scratch.