In theory, the state should have been fixed to...
In theory, the state should have been fixed to work with the new
abstractions for the operations, messaging, and lagrange multipliers.
This is one sort of awkward construction in the constructor, but perhaps
it's not too bad.  Essentially, the state wants to own all the memory
for its members.  Now, in order to make the state manipulator more
flexible, it holds the actual objective function, operations, etc.
Since these elements may allocate memory, we have two options.  Either,
we require a copy constructor so that we can make a copy of what the
user passes in, or we use auto pointers and take conrol of the memory.
Currently, we use auto pointers and take control of the memory.  This is
a little bit awkward since we don't take over the memory of the initial
guess of the variables provided by the user.  Simply, we have an algebra
for actually copying this result, so we do so.  In truth, I don't really
want to add copy constructors to the operations, so I'm fine with this
for now.