I made a pretty substantial change to the...
I made a pretty substantial change to the overall design of the code.
In the last few revisions, I've been playing with passing around a
concrete object that contains all of the algebra for the vector spaces.
In theory, this seems like a good idea, but it causes all sorts of
problems.  Namely, the mathematical abstraction that we're trying to
mimic is a Hilbert space.  In a Hilbert space, we have a vector space
and an inner product.  In a vector space, we have a set, a field,
addition, and scalar multiplication.  Now, the problem occurs because
the set and the field correspond to types.  Since we can't use type
definitions in a concrete object, the set and field became separated
from the operations that defined the Hilbert space.  In short, this
became awkward.  In order to resolve this, we can simply use structures
with all static member functions like we have in the past.  This allows
everything to be packaged together.  Now, the consequence of this and
why we moved away from this design in the first place is that sometimes
the Hilbert space needs to allocate memory for things like the inner
product (sometimes the inner product requires a mass matrix).  In order
to accomplish this with structures that have all static members, we have
to declare some global memory.  At first, I balked at this idea, but in
some sense it does make sense.  Basically, the Hilbert space shouldn't
be changing (the mass matrix shouldn't be changing).  As such, it makes
sense to define it globally once.  Next, I made some substantial changes
to how the optimization state is carried.  I was having trouble trying
to resolve how to use a uniform state for unconstrained, inequality
constrained, equality constrained, and fully constrained problems.  I
tried a number of different tactics such as having elements that were
just ignored or null pointers.  In short, none of these led to
particularly clean code.  In fact, the restart mechanisms capture and
reset became especially arduous.  In order to resolve this issue, I
split the state structure into four separate structures: unconstrained,
equality constrained, inequality contrained, and constrained.
Everything inherits off of unconstrained and constrained inherits off of
equality and inequality constrained.  Anyway, in this manner, it makes
it much easier to manage the individual requirements for inequality or
equality constraints when needed.  Then, in order to work this into the
optimization algorithm, we simply need to template on the state type so
that we only grab what is needed.  In a similar manner, if we template
on the state manipulator, we should have access to all the information
in the state for accomplishing things like restarts.