I pushed back in the quasi-Newton code and...
I pushed back in the quasi-Newton code and this caused a moderate amount
of redesign.  At issue is how we currently interface with the user with
respect to the objective function and preconditioner for the Hessian.
With v3 of the code, we've moved to a design where we abstract user
information about the objective into a new class called
ScalarValuedFunction.  Obstencibly, this is where they implement the
evaluation, gradient, and Hessian-vector product of the objective.  Now,
the question becomes: what is the correct course of action if the user
does not want to provide a Hessian-vector product, but use a
quasi-Newton method.  In the past, the Hessian-vector product was
separated from the objective, so we could just construct one of these
operators if required.  Now, since everything is packed together, as it
probably should be, this becomes more difficult.  The current solution
is to force the user to call a new constructor in any function that
derives off of the ScalarValuedFunction base class.  This constructor
requires both a messaging object and the optimization state.  Based on
the settings in the state, this potentially creates a Hessian-vector
product that overrides the user provided default.  On top of this, we no
longer require the user to implement a Hessian.  As such, if the user
does not implement a Hessian, we can still get the Hessian-vector
product action in a single package.