I added in the packaging for the functions.
I added in the packaging for the functions.  Basically, this means that
there is now a package that contains functions such as the objective
function, preconditioners, or constraints as required.  I futzed around
with this a bit and I finally decided that this packages should own the
memory of the functions.  The reason for this decision is that certain
pieces in this package, such as the preconditioner, may be allocated by
the optimization algorithm.  Therefore, this memory must be owned by the
package unless we want to require the user to muck around with the
details of these preconditions, which I do not.  Also, as a part of
this, it forced a reconsideration of how the inequality constraints are
represented.  At this point, an inequality constraint is simply a
vector-valued function with a line-search.  As far as the cone
definition, I pushed this into the same sort of package as the Hilbert
space.  Basically, all of those operations need to be implemented as
static members of some class along side the other Hilbert space
operations.  This design seemed to make more sense than having a generic
cone object.  Basically, the Euclidean-Jordan algebra operations make
sense to sit along side the Hilbert space operations since these are
operations on elements in a vector space.  The line-search really needs
to sit next to the function definition since how far we can travel in a
particular direction is directly tied to this function.