This is another large change over in the...
This is another large change over in the structure of the code.  In the
previous version, we used the class with static member trick in order to
encapsulate everything we needed in a Hilbert space.  This included both
the operations on elements in this space as well as the vector a field
types.  Using this approach, I started to run into trouble trying to
keep the fields consistent.  Specifically, in equality constrained
problems we have two Hilbert spaces X and Y and each of these has their
own field.  What happens if these fields are different?  The reality is
that this causes all sorts of problems if they are different and are not
compatible.  Frankly, the code probably wouldn't compile.  On top of
this, scalar valued functions return a real type.  This real type may
actually different from the underlying fields in X and Y.  In theory,
everything *should* work if all the spaces are kept consistent.
However, I was having trouble doing so.  For the most part, all my tests
run on doubles, so if I made a mistake, it won't show up.  In order to
debug this, I would have to create incompatible real types and use them
with different Hilbert spaces.  At the same time, this is a fair amount
of work when, really, we want all real types to be the same.  At this
point, I can't honestly think of a good use case where we want to use,
say, a float in one Hilbert space and a doule in another.  Next, I did
explore getting rid of all reals and just using a real field such as a
complex type.  This caused several structural problems as well.
Basically, we really, really need an inner product that induces a norm
and this norm must return a real.  It's difficult to verify this when
the underlying field is complex since the fact that <x,x> is real is
actually a short theorem.  As such, in order to make this work, we'd
have to restrict our inner product to always map to a real.  Certainly,
this is possible and is probably equivalent to an inner product induced
by the real equivalent form, but then we'd have to have every hilbert
space contain both a generic field and a real.  This also probably
calculates our formulations.  Due to all of these issues, I've made the
following design change.  We will continue to work only with real
scalars and not a general field.  Next, we require all Hilbert space
definitions to template themselves on the real type.  By making this
requirement, we can force all associated Hilbert types to use the same
real.  This solves the consistency problems.  This also eliminates some
of the run time checks that I had in order to insure consistency of real
types in certain tests.  Finally, this absolves me from keeping track of
what real type goes where.  At this point, there's just a real.  Of
course, the consequence to this is that the template arguments are
somewhat more complicated.  I hope this isn't too onerous as it greatly
improves the consistency in the code.