I finished the rewrite of the unconstrained...
I finished the rewrite of the unconstrained optimization code.  In
theory, everything should be working correctly now.  Outside of adapting
the old code to the new abstractions, I simplified the nonlinearCG code
so that hopefully it is easier to add different directions.  In
addition, I combined the NewtonCG code and the truncatedCG code.
Basically, they're the same.  Simply, I modify the stopping conditions
depending on whether or not we're line-search or a trust-region.