I made another significant revision and...
I made another significant revision and reorganization.  First, I
changed how the functions are categorized.  Before, I separated things
out between categories such as State, Algorithms, etc.  Then, these were
separated between Unconstrained, EqualityConstraints, etc.  This was ok,
but I kept having to retypedef the types for convenience since the
template parameters are on the Unconstrained, EqualityConstrained, etc.
level.  By moving structures such as Unconstrained to the top level, it
eliminates this problem.  Hence, the difference class of algorithm be it
Unconstrained or Constrained is at the top and then there are
subcategories below it such as State, etc.  Second, I revamped the
restart mechanism by separating it completely from the functions that
support the state.  There are a couple of reasons for this.  Foremost,
there was a ton of code inside the state and it was distracting from
what the state actually contains.  I kept getting confused.  In
addition, there's actually a diamond patern of inheritance within the
states.  Basically, Unconstrained lies at the top and
EqualityConstrained and InequalityConstrained lie at the sides and
Constrained lies at the bottom.  In short, contrained problems contain
elements of both inequality contrained and equality constrained, so the
diamond structure makes sense.  However, this can lead to some problems
with respect to multiple inheritance.  I believe that I resolved these
problems correctly, but I didn't want it to become an issue later, so I
largely removed all functions from the state.  As such, it is mostly
just a dumb structure, which doesn't have any of the same problems with
multiple inheritance.  Third, there were some problems with the restart
mechanism and checking paramters.  Specifically, the code is setup to
throw an error if it doesn't recognize an element during a restart.
Now, the unconstrained code checks all of its parameters by itself, but
the inequality constrained code needs to check both its parameters as
well as the unconstrained parameters.  Since the inequality constrained
optimization contains additional paramters, which are unknown to
unconstrained optimization, this was throwing an error.  The solution to
this is to correctly chain together the verification routines so that an
error is not erroneously thrown.  In the old setup, this was
inconvenient to write.  Hence, the new rewrite resolves these issues.
Now, technically, in constrained optimization, parameters will be
checked more than once due to how things are chained together (both
inequality and equality constrained optimization check the unconstrained
optimization parameters).  I had some code that prevented this, but it
just adds cluster and doesn't hurt anything.  If it becomes an issue
later, the pattern required to accomplish this is used by the code that
translates the restarted lists into the state.  Here, we can not call
the restart code more than once since memory is being moved.