I modified all calculations involving the...
I modified all calculations involving the objective function to instead
use a merit function where appropriate.  The reason for this change is
that I'm trying to bootstrap the inequality constrained optimization
code on top of the unconstrained code and this requires a different
merit function.  Hence, for plain unconstrained optimization, the merit
function is just the objective function.  For constrained optimization,
this is different.  Part of this change is that we no longer cache and
restart with the objective value, but use the merit value instead.