I modified the code so that the predicted...
I modified the code so that the predicted versus actual reduction
calculation uses a generic model function.  We require this modification
so that it's easier to modify the model function when adding inequality
constraints.  Something that looks a little bit odd now, but may be
correct is that normally the constant in the quadratic model function is
the objective value.  Since we no longer store that, I'm using the merit
value.  Now, this seems odd, but it may be correct since we really need
that m(0)=phi(x) where phi is the merit function.  For this to work, we
need the constant to be the merit value.