I completed the initial code for the...
I completed the initial code for the inequality constraints, but it
doesn't completely work properly yet.  The key to this approach is the
design of a state manipulator that allows us to inject the inequality
pieces into the unconstrained code, which together gives an inequality
constrained code.  On top of this, I made some modifications to the
algebra.  Namely, I removed the requirement for a norm function since
that's a possible source of error if sqrt(innr(x,x)) != norm(x).  Also,
I removed the InequalityConstraint sublcass of a vector valued function.
As long as we require the inequality constraints to be affine, I believe
that we can correctly do a line search for feasiblity generically if we
can solve alpha*x+y >=0 for alpha given x and y.  It turns out, the line
search for both the primal and dual variable require a function of this
form.  Finally, I also added the requirement to the vector space for a
barrier function.  This is important for the merit function and it
requires a function barr(x) such that x o grad barr(x) = e.  Hopefully,
with some examples, this is straighforward enough for the user to
implement (all the log barrier stuff works.)  As a final comment, the
code is *not* working.  This is an intermediate check in.