I fixed a bug where I was calculating the...
I fixed a bug where I was calculating the interior point modified
gradient incorrectly.  With this, we're getting the right answer on our
very simple test problem.  The output looks funny (especially the norm
of the gradient, so I need to check that.  I also need to figure out a
way to give some output on the interior point parameter.