I fixed the calculation of the gradient in the...
I fixed the calculation of the gradient in the inequality code.  In the
equations for the optimality conditions, we require the gradient of the
Lagrangian and not the gradient of the objective.  Previously, we just
had the gradient of the objective.  This is not fixed.  In addition,
everytime we reduce our barrier function, we actually change our merit
function and our gradient, so we now recalculate those values.  In
truth, it would be nice if we didn't have to do this, but it does appear
to be necessary otherwise the code seems to get stuck.