I fixed a problem with the unconstrained code...
I fixed a problem with the unconstrained code where we would not
correctly calculate the relative error in truncated-CG when the
algorithm exited after one iteration in Newton-CG.  Most of the time, if
truncated-CG exits during the first iteration of Newton-CG, that means
that negative curvature was detected and we're using the steepest
descent direction.  Hence, the relative error doesn't make any sense.
However, if the RHS is an eigenvector of the the system matrix, we'll
exit after one iteration and the relative error does make sense.  This
has been fixed.  Also, I corrected a problem where the merit function
was incorrect (a sign flipped).  I also corrected the inequality
lagrange multiplier calculation.  At the point in the code where it is
calculated, we've already taken our step in the primal variable.  Hence,
in order to calculate this value correctly, we have to use the previous
iterate.  With these changes, the trust-region code works dramatically
better, but I'm still have issues with the line-search code.