I modified how the interior point parameter in...
I modified how the interior point parameter in the inequality
multipliers are set by default.  In general, neither I nor our users
really know how to pick these quantities.  Typically, I just choose some
random multiplier to have the same norm as the primal variable and then
I set mu to be the estimated multplier.  Generally, this works pretty
well.  However, the question is when to do this especially given that
the initial state may come from a restart.  Now, I take the following
approach.  By default, mu is now NaN.  If it is NaN when it hits the
optimization, I set it to be the estimated parameter.  Otherwise, I let
it through.  For the inequality multiplier z, during its initialization,
I set it to be norm(x)*e.  In the case of a restart, this routine should
not be called.  Also, I changed the example to use BFGS.  Happily, it
works with inequality constraints.