I fixed some minor bugs in peopt and I added a...
I fixed some minor bugs in peopt and I added a number of different
driver routines for doing cone programming.  Basically, for single
computer multithreaded machines, I added all of the vector spaces
required for semidefinite, quadratic, and linear programming.  On top of
this, I added some additional examples to insure that everything is
working correctly.  I still need to optimize a few things.  Right now,
the Linv operator semidefinite programming is expensive to compute since
it requires the solution to the Sylvester equations and we accomplish
this through a Schur decomposition.  Perhaps there's a better way to
accomplish this, but that's the case for now.  Anyway, it's very
important that we don't over compute the Schur decomposition, but cache
it.  I've added the caching functionality, but the core routines need to
be modified so that this is not computed so often.  In addition, the
barrier function for SDP also uses the Schur decomposition.  I did this
because I thought that the values were going to be cached, but I'm
starting to have doubts since the line search routines call the barrier
function very often at points in the line search.  As such, it's
probably best to compute this value based on the Choleski factorization.
In anycase, everything for the most part is there and appears to work.