I completely re-did the diagnostics so that it...
I completely re-did the diagnostics so that it was easier to actually
output information about the interior point method.  Previously, the
diagnostic tests were hard coded into the unconstrained optimization
code.  This became inconvenient when working with the inequality
constrained code since we essentially bolt on an interior point method.
In the process of bolting on the interior point method, we did not have
diagnostics displayed regarding the interior point parameters.  In order
to fix this, I made a few changes.  First, all of the hard coded
diagnostics have been removed and replaced with a new system that uses
the state manipulator.  Next, the process of assembling diagnostic
information has changed.  Now, I compile a list of things that may be
useful to output and pass the list around.  Only at the very last step
is this list turned into a printable string.  This allows us to shove
new diagnostic information into the list when required, which is the
case with the interior point method.  As a consequence, our output isn't
quite as tight as it was before.  Basically, every output takes exactly
10 characters, which is too much for some and too little for others.
However, it does make everything easy to parse out later, since we can
get all the diagnostics for a single variable by just parsing the output
through the unix cut command.  In addition, the varying print levels
have changed slightly.  Now, every parameter has it's own column, which
include the extra information about, say, the Krylov method.  On
iterations where this information is not required, we print blank
information.  Outside of that, I also simplified the messaging object
since I was a little perturbed with remembering what the level meant.
Hopefully, this makes it easier for the user to implement their own
messaging device.  I also fixed a bug with the interior point stopping
condition where we were not stopping when we should have.  Basically, I
was overriding the default stopping condition when I should not have.
In addition, we had a situation where the trust-region loop could enter
an infinite loop when the trust-region radius became too small.  I added
an extra check that forces an exit when the step during this subproblem
becomes too small according to the parameter eps_s.  I also put this
check into the line-search code as there is a risk that something
similar occurs when doing a line-search.  In theory, this should never
happen, but there appeared to be cases where the size of the step
compared to the size of the iterate was below machine precision, so we
can't accurately calculate the new objective value or new step.