I fixed a small bug with the debugging output...
I fixed a small bug with the debugging output where the columns did not
align.  I also added a JSON parser and an addon.  Basically, I don't
really think it's appropriate to have the JSON parser as part of the
core utilities since the user may have some odd way to parse.  However,
it's useful, so I added one.