I eliminated the need to store a messaging...
I eliminated the need to store a messaging object inside a SQL vector.
Basically, the only time an error would be thrown is during the initial
construction of the vector.  After that, as long as we copy from a valid
vector, everything should work.  Therefore, we pass in a messaging
object during the construction, but do not require it when copying or
any other operation.