I modified how the state manipulators are created.
I modified how the state manipulators are created.  Previously, both the
state and the function types were passed as template parameters.  This
has changed.  Now, we pass in a single problem class, which must contain
state and function types as members.  In theory, this more closely
represents what is being done.  In addition, I added a conversion
manipulator that converts from one problem class to another.  Basically,
the unconstrained code requires an unconstrained state manipulator.
When we call the interior point code, we pass in an inequality
constrained state, which is a subclass of the unconstrained state.  As
such, the unconstrained code should interpret this code just fine.
However, in order to run the inequality constrained code correctly, we
need an inequality state manipulator to insert pieces where needed.
This violates the requirement for a unconstrained state manipulator.
Now, we could just template the unconstrained code on the state and
manipulator so that there's no conflict, but I don't particularly want
to keep track of those parameters.  Alternatively, we could invert the
design of the code so that all of these functions are held internal to
the object and let polymorphism do it's job.  Although this should work,
it makes the rest of the design inconvenient, so I don't really want to
pursue this either.  Frankly, it was just easy to dynamically cast the
state since we know it's a safe cast and it's not going to affect the
computational speed.  Hence, the conversion manipulator just casts over
the types to those specified.