I modified and extended how we can setup a...
I modified and extended how we can setup a valid optimization state.
Previously, all initialization occured through the constructors.
Although this worked, it made some operations awkward.  For example,
sometimes we just want to initialize a state and then examine the
nonvector parameters.  This helps tell the user the defaults.  However,
if we do this, we didn't have a good mechanism for inserting the vectors
later.  Basically, we'd have to make a dummy state to examine for
parameters and then construct a new state that has all the vectors
initialized.  This has been fixed.  In theory, we should be able to use
the constructors as before if we so choose.  In addition, I eliminated
the redundant unconstrained parameter check.  It didn't really hurt
anything, but we might as well not do it twice.