I corrected a bug where the norm of s and...
I corrected a bug where the norm of s and s_typ were not restarted
correctly.  In addition, I set the initial value of the krylov_rel_err
to zero from nan.  Basically, if we're not using a krylov method and we
restart, having a nan in the position is inconvenient since json doesn't
know how to read it in properly.  In some sense, we don't really need
this value unless the krylov method is on, so setting it to zero
shouldn't hurt antyhing.