I don't think this entirely matters, but I...
I don't think this entirely matters, but I changed how the initial
lagrange multiplier for the inequality constrained problem is set.
Basically, I figure that a lagrange multiplier with the same norm as x
should be an ok starting point, so I wanted to set a the paramter where
||z|| = || x || and z is some multiple of the jordan identity element.
That wasn't happening, so I fixed that problem.