I fixed an inefficiency with the last check-in,...
I fixed an inefficiency with the last check-in, which introduced a bug,
which is fixed with this check-in.  Basically, we had been
overallocating memory for the caching of the Schur-decomposition.  This
had been fixed, but not all of the offset information was updated
elsewhere in the code.  Hopefully, this has been recitified now.