There was a subtle error in the formulation...
There was a subtle error in the formulation that led to the interior
point modification of the Hessian to be nonsymmetric in the case of SDP
constraints.  In order to fix this, we now forcefully symmetrize this
piece.  We should be in the clear theoretically, since modify the
objective with the same quadratic function.  Merely, we use the
symmetric quadratic form now rather than a nonsymmetric form.  Now,
technically, this fix is not needed if we use linear inequality
constraints.  However, from what I can tell, the computational cost of
the symmetrization, where we do it, is almost nil, so I don't think it
will affect computational performance.