I fixed a bug where a pair of quasi-Newton...
I fixed a bug where a pair of quasi-Newton vectors was added to the
stored history when BFGS is selected and their inner product is zero.
This inner product should be positive.  In theory, if we checked the
strong-Wolfe conditions, we could verify this fact.  However, we don't,
so we monitor the situation directly.