I made several modifications and fixes.
I made several modifications and fixes.  First, I modified the build
process to hopefully make things simpler.  In the most basic case, we
only require the header, so anything that only links into the core
routines doesn't have to be changed.  However, if someone wants to use
the default vector spaces, everything now compiles to a statically
linked library.  In order to setup the build, copy make.inc.example to
make.inc and adjust the variables.  Second, I believe there to be a bug
in the lapack routine dstemr.  Basically, the eigenvalues are supposed
to be returned in ascending order, but this doesn't appear to be the
case for matrices of size 2.  This interferes with out Lanczos code,
which requires a dense solve of small matrices in order to calculate the
Ritz values.  I've reported the error to the lapack developers, but
until we can verify the problem and obtain a fix, I've changed the
underlying eigenvalue routine to dstevr.  Of course, I believe dstemr
calls dstevr, but the eigenvalues appear to be correctly sorted in
dstevr.  Third, I added the start of some basic unit tests in the unit
directory.  These need to be expanded upon, but it provides some initial
framework for testing.  It's also probably good to compile them
regularly during development since C++ doesn't catch some errors with
templated code unless the routines are instantiated.