This is a large change over in the peopt code...
This is a large change over in the peopt code where we now use
autoconfig in order to build the system.  In the simplest case, we just
need the headers, so we don't really compile anything.  In order to make
this work, we simply do a:  ./configure --with-blas=no --with-boost=no
--prefix=<stuff>.  If we want access to additional routines, the new
build procedure is documented in README.