This new build script appears to work with...
This new build script appears to work with Macs that do not include a
Fortran compiler.  Basically, we do a hard check for the compiler and if
we don't find it, we do not attempt to search for BLAS or LAPACK.  Then,
this means that we not only don't build the predefined vector spaces,
but we don't install the headers for this either.