It turns out that the previous scheme did not...
It turns out that the previous scheme did not work on all our platforms.
Basically, some platforms will suggest that they have a working Fortran
compiler, but this may not be the case.  As such, either we need to
rewrite some standard autoconfig scripts (not recommended), or we need
to forcefully shortcut some of the configuration.  I've taken the later
route and added a flag that should do some of this shortcircuiting.  It
needs some testing, but it may work.