For some reason, I missed a change I had made...
For some reason, I missed a change I had made earlier to only install
two headers for the headers-only build.  This has been fixed.
Basically, the other two headers deal with a predefined set of linear
algebra.  If we're not compiling their definitions, there's not a good
reason to install the headers.