I added the infrastructure for constructing,...
I added the infrastructure for constructing, restarting, validating, and
initializing equality constrained optimization problems.  Something that
surprised me a little bit was the preconditioner for the augmented
system.  In theory, our code should be able to accept an arbitrary
precondioner for this system.  Howevever, the precondioner would need to
work on the space X x Y, which is a little messy to code since I'd have
to define product spaces.  In reality, we're probably going to use a
Schur complement based precondioner, which only requires an operator on
the space Y.  This fits more naturally into our design.  Perhaps later,
it'll be worth defining general preconditoners, but we'll ignore that
for now.