I moved the operator definition to the linear...
I moved the operator definition to the linear algebra and added a
dependence on peopt to the linear algebra.  This means that we'll no
longer have a pure header library since we depend on BLAS for GMRES.
Technically, I've reimplemented the required routines in a purely
templated fashion, but, I'll be honest, I don't to maintain those files
nor be responsible for their maintainene.  In any case, I've also added
a fully general GMRES.  Technically, this doesn't have everything we
need since I need to add a capacity for stopping conditions that vary at
every iteration.  I think I can play a trick similar to the state
manipulator in order to implement this.