I modified GMRES to allow flexible stopping...
I modified GMRES to allow flexible stopping conditions.  This required
some restructuring so that we do not apply the operator or the
preconditioner more than we require since they could be extremely
expensive to compute.  As far as the flexible stopping conditions, they
really only allow an absolute tolerance, so it makes this routine less
useful in general, but appropriate for Denis and Matthias' algorithm.