I added unit tests for GMRES and fixed the...
I added unit tests for GMRES and fixed the build accordingly.  In
addition, I fixed some minor bugs in the GMRES routine.  Namely, I was
not calculating the new iterate correctly as I was using more Krylov
vectors than the current iterate.  Recall, at iteration m, we have m+1
Krylov vectors.  However, when we do a backsolve for the new iterate, we
have a mxm system.  It's important not to use this last vector since it
may be zero if breakdown occurs.  Since we normalize the Krylov vectors
before storing them, this can create a Krylov vector with NaNs in it,
which causes problems if we use it.