I added a new parameter to control the number...
I added a new parameter to control the number of GMRES (or other
iterative method) iterations for the augmented system.  This is a little
confusing since there's another parameter krylov_iter_max, which deals
with the tangential subproblem in the equality constrained algorithm or
the Newton system in the unconstrained algorithm.  This parameter should
probably change name.  When I can think of a better one, I'll do that.