I added memory allocation and restarting to...
I added memory allocation and restarting to some new vectors for the
equality constrained problems.  These include the tangential step,
normal step, cauchy point, and the projected tangential step.  This
turned out to be a little tricky since we have a new situation with
memory allocation.  Before, the only X vectors allocated were in the
unconstrained code.  Similarly, we only allocated Y vectors in the
equality constrained code and Z vectors in the inequality constrained
code.  Now, we have to allocate both X and Y vectors in the equality
constrained code.  This required a revamp of the restart utilities since
they were not setup to account for multiple algorithms capturing the
state.  Specifically, when we do a capture of the state, the
unconstrained restart routines need to capture the memory for several X
vectors, but not all of them.  The remaining X vectors need to be
captured by the equality constrained code.  This is a little tricky
since the capture routine is destructive by nature.  Hence, we needed to
be much more clean when we remove elements from the lists.  In theory,
the new setup should take care of those details.